Introduce ScopedCell and ScopedRefCell
As part of a plan to overhaul the "scoped_set" pattern, introduce two new
"scoped" types:

ScopedCell is a trivial wrapper around Cell which exposes a function
`scoped_mod`. This allows you to modify the contents of the Cell; the
original contents are restored when the guard is dropped.

ScopedRefCell is similar, except it requires the caller to supply an
accessor which returns a &mut field, and also the new value.

These will be used to replace the rather unsightly scoped_push calls
around the fish codebase.