Remove intrusive Screen::scoped_buffer
scoped_buffer borrows the screen to be able to flush buffers on drop.  This is
a bit intrusive (see also 29ae571afa (Make scoped_push nicer, 2024-12-28)).

Not sure what's the best solution -- probably we should always pass the
outputter as parameter, removing the field.

For now let's at least contain the effects of this.