Stop supporting terminals without eat_newline_glitch
This reverts commit 58347d494a (update PROMPT_SP heuristic, 2016-12-23)

If we write the last column of a line:

	printf %0"$COLUMNS"d 0; sleep 3

most terminals will *not* move the cursor to the next line.

This behavior is indicated by the terminfo's xenl (AKA xen or
eat_newline_glitch) boolean capability.

We originally added checks for this capability because ConEmu and Windows
ConHost did not implement it, but they do now.

Also, as mentioned in
https://github.com/fish-shell/fish-shell/pull/11342#issuecomment-2769979520,
we're not aware of any present-day terminal that does not have this behavior,
but the value advertised by terminfo is sometimes wrong.

Let's get rid of this for now. A following commit will document that we
require this behavior.