Outputter cache to use text face type
When the text face changes from bold to non-bold, we need to reset
all attributes since there is no exit_bold_mode.

The same holds for various dim and reverse mode, and future attributes.

A following commit wants to make this logic more robust to addition of future
attributes. Prepare for that by using the new type.