Remove unused query for cursor position reporting
We only query for the cursor position if either
1. if the terminal sends an XTGETTCAP response indicating it supports scroll
   forward (terminfo code indn)
2. if the terminal sends a mouse click event

In practice, the terminals that do either are also capable of reporting the
cursor position, so let's require that for now.

We could definitely make this optional, and degrade gracefully -- that's
what I originally intended.

But maybe we don't need it? Keeps things simpler. In future, we can add a
timeout, and print an error, to help debugging terminals.