Prevent code execution from focus events while blocked on query response
While we are waiting for a query response from the terminal, we defer any
input processing until we receive our query response
 Then that response is promoted to the front of the input queue, and
remaining inputs are processed in order.

We accidentally process focus events, which may run arbitrary code.  We can't
do this; it breaks a lot of invariants (for example the it can invalidate
the cursor positions for CursorPositionQuery, or it can cause fish to exit
before fully consumeing a query response).

Make sure we only process known-safe events.  We definitely need to process
CheckExit (in case we received SIGHUP). I guess we should also process Eof,
in case the terminal is buggy.