ast: remove the ability to traverse in reverse
Prior to this commit, each Node in the ast could accept a visitor and visit
children either in order or in reverse order. This reverse feature added a lot
of complexity and the only client is Traversal.

Switch Traversal to reverse the nodes itself and remove the reverse bool
elsewhere, leading to some code simplifications.