Fix diff_profiles.fish
This script was broken by the changes to profiling output in
9d904e111345a64eb4ff05ecd8d7737f22cfc762.

The new version works with both the old and new profiling output, even when
mixed. The script output has been adjusted to match the new profiling style
better.

This also adds basic error handling for situations where the script is invoked
incorrectly and makes the file executable.