Add support for using embedded mo file
Embedding the MO files for gettext means that we cannot call the system gettext
anymore, since there is no actual file it can read from (unless we write out the
embedded data onto the file system).

Therefore, the `gettext` crate is used. This crate does not seem to be
maintained, but it works well enough for this proof of concept.

For this to work, the MO files need to be present at build time. Run
`build_tools/update_translations.fish` to ensure that this is the case.
Binaries built this way have the MO files embedded (if the `embed-data`) feature
is active. Setting the relevant environment variables should then result in
translations working as expected for the most part.

See https://www.gnu.org/software/gettext/manual/gettext.html#Locale-Environment-Variables-1
for the supported variables. My implementation roughly follows the concepts
described there, but there are probably some differences to the original gettext
implementation, especially for the more complex cases.
Fallback logic is not ideal at the moment. E.g. if someone sets `LANGUAGE` to
`pt`, for which no translations exist, the fallback to use the source strings,
instead of `pt_BR` which does exist. The other way around (from language with
region to just language) works.

The code is quite messy at the moment, as this is just intended as a
proof-of-concept.