Use separate lock file for history file accesses
Locking on the history file itself comes with several problems related to the
file being renamed in the `save_internal_via_rewrite` function.
Specifically, before locking a file, it needs to be opened.
It is possible that the file gets replaced after opening but before locking.
The code had checks for this using the inode number and timestamps, but even
with these present, races could occur.

The simpler approach is to use a separate file for locking. This file never
needs to be replaced, so the aforementioned problem does not occur.

This version of the code provides a struct whose constructor allows opening the
history file while holding a lock on it. This lock can be either exclusive or
shared.
The constructor opens (and possibly creates) the lock file and locks it.
Once the lock is held, the history file is also opened.
A struct holding both files is returned.
At the moment, the encapsulation is weak, since both files are accessible to the
caller of the constructor, even though there is no need to access the lock file.
The lock on it is released when the `File` is dropped. (Because then, the file
descriptor is closed.)
Directly exposing the history file is also questionable, but convenient for
keeping compatibility with existing code.

For now, HistoryImpl::{populate_from_config_path, clear} access the history file
without locking. This is unchanged from the previous behavior.

One important additional change is that the rewrite path for updating the
history file now takes the lock before reading the history file, which is done
before populating the temporary file. This is necessary and not doing so
introduced a race condition in the old code.