Do not abandon locking if it took a long time
Abandoning locking can cause corruption of the history data. Especially in
situations where locking takes a long time, it is likely to be useful. That is,
for scenarios with a lot of contention (which probably mostly happens in
tests/benchmarks), and if accesses take a long time due to a remote file system.

Maybe we can add a user-configurable setting to disable locking, but I think it
is a bad idea to do so by default if locking took a bit longer than desirable
once.