history: remove flush()/fsync() and don't write after each appended item
Commit 5c0fddae70a (Refactor history flushing, 2025-03-28) made three changes:
1. call fsync() when we are finished writing the history file. 
2. when appending to (as opposed to vacuuming) history, call write(2) (+
   flush()+sync()) for each item. Previously, we'd only call write(2) if our
   64k buffer was full, or after processing the last history item.
3. actually check the return value of flush() (which would retry when flushing
   fails -- but std::fs::File::flush() never fails).

The motivation was to potentially fix #10300 which didn't succeed (see
https://github.com/fish-shell/fish-shell/issues/10300#issuecomment-2876718382).
I also don't understand how it can fix that issue.

This part of the code is very tricky so we should avoid muddying the waters.
We should avoid merging changes that dont' have a concrete, explicitly
documented motivation.

I don't understand yet what is the benefit of 1 and 2.  If there is one,
I'd love to learn so I can use it in future.

As for 3, there are some explanations in comments, commit message and a [PR
comment](https://github.com/fish-shell/fish-shell/pull/11330#discussion_r2020171339).
To summarize those, 5c0fddae70a wants to address the scenario where
file.flush() fails. Prior to that commit we would ostensibly carry on with
a corrupted "first_unwritten_new_item_index" (corrupted because it doesn't
match what's written to disk), which can cause various issues.
However this doesn't ever happen because std::fs::File::flush() never fails --
File does not buffer writes, it always passes through writes to write(2).

There can definitely be scenarios like the one described in
https://github.com/fish-shell/fish-shell/pull/11330#discussion_r2020171339
where the disk is full. In that case, either write(2) fails, which we
already check.  Or close(3p) fails with EIO, which we have never checked. We
should probably check that, I'm happy to add that fix and test it.

Undo all three changes for now.
I think a good way forward would be for a user who experiences #10300 to
create a reproducer by running `history append` in a loop.