Add error handling for history-related functions
The main goal of this is to have more meaningful error messages for
history-related functionality. In some cases this allows detecting errors where
it was not possible before.

Some non-history related code which calls the same dependencies is also updated.
The largest impact is on the functionality for creating temporary files.
There were two separate functions which basically did the same thing, which are
now merged into one. They retried unconditionally when creating a temporary file
failed. In most situations this does not make sense, so we explicitly check for
cases where retrying immediately might work (i.e. when the process was
interrupted), and only retry in such cases.